pyproject.toml
[project]
name = "elt-dataflow-n8n"
version = "0.1.0"
description = ""
authors = [
    {name = "ValmirJrDS",email = "valmirmoreirajunior@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "dbt-postgres (>=1.9.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
